// frontend/src/api/models/AreaItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AreaItem = {
    id?: string;
    name?: string;
    ship_count?: number;
    code?: string;
    timezone?: string;
    org_id?: string | null;
};